Rearranged a bit. Put events in one place for now.
