+++ Rpg.Mobile.App/Game/MainBattle/MainBattleStateMachine.cs
@@ -1,4 +1,5 @@
 ﻿using Rpg.Mobile.App.Game.MainBattle.Components;
+using Rpg.Mobile.App.Game.MainBattle.Events;
 using Rpg.Mobile.App.Game.MainBattle.States;
 using Rpg.Mobile.App.Game.MainBattle.Systems.Calculators;
 using Rpg.Mobile.App.Game.MainBattle.Systems.Data;
@@ -107,8 +108,4 @@ public class MainBattleStateMachine : StateMachine<IBattleState>
 
 public interface IBattleState : IState { }
 
-public record UnitTurnEndedEvent(BattleUnitData Unit) : IEvent;
-public record UnitDamageAssignedEvent(IEnumerable<BattleUnitData> Units, int Damage) : IEvent;
-public record BackClickedEvent : IEvent;
-public record AttackClickedEvent : IEvent;
-public record MagicClickedEvent : IEvent;
+
